add command and behaviors

+++ EventService.API/DTOs/Event/CreateEventRequest.cs
@@ -3,9 +3,9 @@ public record CreateEventRequest
 (
     string Name,
     string Image,
-    long NoVoucher,
+    int NoVoucher,
     DateTime StartDate,
     DateTime EndDate,
-    long BrandId,
-    long? GameId
+    int BrandId,
+    int? GameId
 );
